Basic changes.  Addition of ISI triggered histogram filter, which outputs a spatio-temporal histogram that starts when volume exceeds some level.
git-svn-id: https://svn.code.sf.net/p/jaer/code/trunk/host/java@3503 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ src/ch/unizh/ini/jaer/projects/neuralnets/ISIMapper.java
@@ -35,7 +35,7 @@ public class ISIMapper extends NetMapper{
     
 //    private float minFreqHz;
 //    private float maxFreqHz;
-    public float nBins;
+    public int nBins;
     
     
     public ISIMapper(int nChannels,int nBin)
